feat(ParticleEmitter): 	- add a particle emitter component that alows for particle 	  generation 		- particles can be rendered via batching or newly added 		  instancing depending on user preference 		- particles may use additive blending instead of 		  1 - src alpha blending

+++ engine/core/engine/Buffer.hpp
@@ -19,7 +19,7 @@ namespace s2f
 
 		GLuint id() const { return mID; }
 		BufferType type() const { return mType; }
-		u32 elementCount() const { return mElementCount; }
+		u64 elementCount() const { return mElementCount; }
 
 		bool valid() const;
 		void setData(u64 size, void* data) const;
